refactor: move new changes to rfc7523
