+++ handler/oauth2/storage.go
@@ -23,10 +23,7 @@ package oauth2
 
 import (
 	"context"
-	"time"
-
 	"github.com/ory/fosite"
-	"gopkg.in/square/go-jose.v2"
 )
 
 type CoreStorage interface {
@@ -67,26 +64,4 @@ type RefreshTokenStorage interface {
 	GetRefreshTokenSession(ctx context.Context, signature string, session fosite.Session) (request fosite.Requester, err error)
 
 	DeleteRefreshTokenSession(ctx context.Context, signature string) (err error)
-}
-
-// JWTAuthGrantStorage holds information needed to validate jwt assertion in authorization grants.
-type JWTAuthGrantStorage interface {
-	// GetPublicKey returns public key, issued by 'issuer', and assigned for subject. Public key is used to check
-	// signature of jwt assertion in authorization grants.
-	GetPublicKey(ctx context.Context, issuer string, subject string, keyId string) (*jose.JSONWebKey, error)
-
-	// GetPublicKeys returns public key, set issued by 'issuer', and assigned for subject.
-	GetPublicKeys(ctx context.Context, issuer string, subject string) (*jose.JSONWebKeySet, error)
-
-	// GetPublicKeyScopes returns assigned scope for assertion, identified by public key, issued by 'issuer'.
-	GetPublicKeyScopes(ctx context.Context, issuer string, subject string, keyId string) ([]string, error)
-
-	// IsJWTUsed returns true, if JWT is not known yet or it can not be considered valid, because it must be already
-	// expired.
-	IsJWTUsed(ctx context.Context, jti string) (bool, error)
-
-	// MarkJWTUsedForTime marks JWT as used for a time passed in exp parameter. This helps ensure that JWTs are not
-	// replayed by maintaining the set of used "jti" values for the length of time for which the JWT would be
-	// considered valid based on the applicable "exp" instant. (https://tools.ietf.org/html/rfc7523#section-3)
-	MarkJWTUsedForTime(ctx context.Context, jti string, exp time.Time) error
 }
