+++ handler/rfc7523/flow_authorize_jwt_session.go
@@ -19,8 +19,10 @@
  *
  */
 
-package oauth2
+package rfc7523
 
-type AuthorizeJWTGrantStorage interface {
-	JWTAuthGrantStorage
+// AuthorizeJWTGrantSession must be implemented by the session if RFC7523 is to be supported.
+type AuthorizeJWTGrantSession interface {
+	// SetSubject sets the session's subject.
+	SetSubject(subject string)
 }
